Removed ability to delete user character not related to current user

+++ Progress.Application/Usecases/UserCharacters/Delete/DeleteUserCharacterCommand.cs
@@ -3,6 +3,7 @@ using LanguageExt;
 using MediatR;
 using Progress.Application.Common;
 using Progress.Application.Persistence;
+using Progress.Application.Security.Services;
 
 namespace Progress.Application.Usecases.UserCharacters.Delete
 {
@@ -16,17 +17,26 @@ namespace Progress.Application.Usecases.UserCharacters.Delete
     public class DeleteUserCharacterCommandHandler : ValidationRequestHandler<DeleteUserCharacterCommand, Unit>
     {
         private readonly ApplicationDbContext dbContext;
+        private readonly ICurrentUser currentUser;
 
-        public DeleteUserCharacterCommandHandler(ApplicationDbContext dbContext, IEnumerable<IValidator<DeleteUserCharacterCommand>> validators) : base(validators)
+        public DeleteUserCharacterCommandHandler(ApplicationDbContext dbContext, ICurrentUser currentUser,
+            IEnumerable<IValidator<DeleteUserCharacterCommand>> validators) : base(validators)
         {
             this.dbContext = dbContext;
+            this.currentUser = currentUser;
         }
 
 
-        protected override async Task<Either<Failure, Unit>> WrappedHandle(DeleteUserCharacterCommand request, CancellationToken cancellationToken)
+        protected override async Task<Either<Failure, Unit>> WrappedHandle(DeleteUserCharacterCommand request,
+            CancellationToken cancellationToken)
         {
-            var entityToRemove = dbContext.UserCharacters.Single(uc => uc.Id == request.Id);
+            var entityToRemove = dbContext.UserCharacters.SingleOrDefault(uc => uc.Id == request.Id && uc.UserId == currentUser.Id);
 
+            if (entityToRemove is null)
+            {
+                return new Failure(new Exception("Error when trying to delete specified user character"));
+            }
+            
             dbContext.UserCharacters.Remove(entityToRemove);
 
             await dbContext.SaveChangesAsync(cancellationToken);
